move config validation logic
