+++ cs2tracker/validated_config.py
@@ -1,0 +1,74 @@
+from configparser import ConfigParser
+
+from cs2tracker.constants import CAPSULE_INFO, CONFIG_FILE
+from cs2tracker.padded_console import PaddedConsole
+
+console = PaddedConsole()
+
+
+class ValidatedConfig(ConfigParser):
+    def __init__(self):
+        """Initialize the ValidatedConfig class."""
+        super().__init__(interpolation=None)
+        super().read(CONFIG_FILE)
+
+        try:
+            self._validate_config()
+            self.valid = True
+        except ValueError as error:
+            console.print(f"[bold red][!] Config error: {error}")
+            self.valid = False
+
+    def _validate_config_sections(self):
+        """Validate that the configuration file has all required sections."""
+        if not self.has_section("User Settings"):
+            raise ValueError("Missing 'User Settings' section in the configuration file.")
+        if not self.has_section("App Settings"):
+            raise ValueError("Missing 'App Settings' section in the configuration file.")
+        if not self.has_section("Custom Items"):
+            raise ValueError("Missing 'Custom Items' section in the configuration file.")
+        if not self.has_section("Cases"):
+            raise ValueError("Missing 'Cases' section in the configuration file.")
+        for capsule_section in CAPSULE_INFO:
+            if not self.has_section(capsule_section):
+                raise ValueError(f"Missing '{capsule_section}' section in the configuration file.")
+
+    def _validate_config_values(self):
+        """Validate that the configuration file has valid values for all sections."""
+        try:
+            for custom_item_name, custom_item_owned in self.items("Custom Items"):
+                if " " not in custom_item_owned:
+                    raise ValueError(
+                        f"Invalid custom item format (<item_name> = <owned_count> <item_url>): {custom_item_name} = {custom_item_owned}"
+                    )
+                owned, _ = custom_item_owned.split(" ", 1)
+                if int(owned) < 0:
+                    raise ValueError(
+                        f"Invalid value in 'Custom Items' section: {custom_item_name} = {custom_item_owned}"
+                    )
+            for case_name, case_owned in self.items("Cases"):
+                if int(case_owned) < 0:
+                    raise ValueError(
+                        f"Invalid value in 'Cases' section: {case_name} = {case_owned}"
+                    )
+            for capsule_section in CAPSULE_INFO:
+                for capsule_name, capsule_owned in self.items(capsule_section):
+                    if int(capsule_owned) < 0:
+                        raise ValueError(
+                            f"Invalid value in '{capsule_section}' section: {capsule_name} = {capsule_owned}"
+                        )
+        except ValueError as error:
+            if "Invalid " in str(error):
+                raise
+            raise ValueError("Invalid value type. All values must be integers.") from error
+
+    def _validate_config(self):
+        """
+        Validate the configuration file to ensure all required sections exist with the
+        right values.
+
+        :raises ValueError: If any required section is missing or if any value is
+            invalid.
+        """
+        self._validate_config_sections()
+        self._validate_config_values()
